start bilding functionality. todo: students, students_* api endpoints.
