+++ app/Http/Controllers/Api/V1/Auth/EmailVerificationController.php
