+++ app/Http/Resources/Api/V1/PostResource.php
